Primera version casi lista

+++ composeApp/src/commonMain/kotlin/com/easypark/app/parkingdetails/presentation/state/ParkingDetailsEvent.kt
@@ -4,5 +4,5 @@ sealed interface ParkingDetailsEvent {
     object OnBackClick : ParkingDetailsEvent
     object OnReserveClick : ParkingDetailsEvent
     data object OnLoadDetail : ParkingDetailsEvent
-    data class OnRate(val rating: Float) : ParkingDetailsEvent
+    data class OnRate(val rating: Int) : ParkingDetailsEvent
 }
